Update my master

+++ configmgr/source/components.cxx
@@ -21,8 +21,9 @@
 
 #include <algorithm>
 #include <cassert>
+#include <chrono>
 #include <cstddef>
-#include <list>
+#include <vector>
 #include <set>
 
 #include <com/sun/star/beans/Optional.hpp>
@@ -30,12 +31,14 @@
 #include <com/sun/star/beans/XPropertySet.hpp>
 #include <com/sun/star/container/NoSuchElementException.hpp>
 #include <com/sun/star/lang/WrappedTargetException.hpp>
+#include <com/sun/star/lang/WrappedTargetRuntimeException.hpp>
 #include <com/sun/star/uno/Any.hxx>
 #include <com/sun/star/uno/Exception.hpp>
 #include <com/sun/star/uno/Reference.hxx>
 #include <com/sun/star/uno/RuntimeException.hpp>
 #include <com/sun/star/uno/XComponentContext.hpp>
 #include <com/sun/star/uno/XInterface.hpp>
+#include <cppuhelper/exc_hlp.hxx>
 #include <config_dconf.h>
 #include <config_folders.h>
 #include <osl/conditn.hxx>
@@ -45,10 +48,10 @@
 #include <rtl/ref.hxx>
 #include <rtl/ustrbuf.hxx>
 #include <rtl/ustring.hxx>
-#include <rtl/instance.hxx>
 #include <sal/log.hxx>
 #include <sal/types.h>
 #include <salhelper/thread.hxx>
+#include <comphelper/backupfilehelper.hxx>
 
 #include "additions.hxx"
 #include "components.hxx"
@@ -66,7 +69,7 @@
 #include "xcsparser.hxx"
 
 #if ENABLE_DCONF
-#include <dconf.hxx>
+#include "dconf.hxx"
 #endif
 
 #if defined(_WIN32)
@@ -77,17 +80,17 @@ namespace configmgr {
 
 namespace {
 
-struct UnresolvedListItem {
+struct UnresolvedVectorItem {
     OUString name;
     rtl::Reference< ParseManager > manager;
 
-    UnresolvedListItem(
+    UnresolvedVectorItem(
         OUString const & theName,
-        rtl::Reference< ParseManager > theManager):
+        rtl::Reference< ParseManager > const & theManager):
         name(theName), manager(theManager) {}
 };
 
-typedef std::list< UnresolvedListItem > UnresolvedList;
+typedef std::vector< UnresolvedVectorItem > UnresolvedVector;
 
 void parseXcsFile(
     OUString const & url, int layer, Data & data, Partial const * partial,
@@ -128,10 +131,9 @@ bool canRemoveFromLayer(int layer, rtl::Reference< Node > const & node) {
     switch (node->kind()) {
     case Node::KIND_LOCALIZED_PROPERTY:
     case Node::KIND_GROUP:
-        for (NodeMap::const_iterator i(node->getMembers().begin());
-             i != node->getMembers().end(); ++i)
+        for (auto const& member : node->getMembers())
         {
-            if (!canRemoveFromLayer(layer, i->second)) {
+            if (!canRemoveFromLayer(layer, member.second)) {
                 return false;
             }
         }
@@ -154,7 +156,7 @@ public:
     void flush() { delay_.set(); }
 
 private:
-    virtual ~WriteThread() {}
+    virtual ~WriteThread() override {}
 
     virtual void execute() override;
 
@@ -177,17 +179,14 @@ Components::WriteThread::WriteThread(
 }
 
 void Components::WriteThread::execute() {
-    TimeValue t = { 1, 0 }; // 1 sec
-    delay_.wait(&t); // must not throw; result_error is harmless and ignored
+    delay_.wait(std::chrono::seconds(1)); // must not throw; result_error is harmless and ignored
     osl::MutexGuard g(*lock_); // must not throw
     try {
         try {
             writeModFile(components_, url_, data_);
         } catch (css::uno::RuntimeException & e) {
             // Ignore write errors, instead of aborting:
-            SAL_WARN(
-                "configmgr",
-                "error writing modifications: \"" << e.Message << '"');
+            SAL_WARN("configmgr", "error writing modifications: " << e);
         }
     } catch (...) {
         reference_->clear();
@@ -196,19 +195,12 @@ void Components::WriteThread::execute() {
     reference_->clear();
 }
 
-class theComponentsSingleton :
-    public rtl::StaticWithArg<
-        Components,
-        css::uno::Reference< css::uno::XComponentContext >,
-        theComponentsSingleton>
-{
-};
-
 Components & Components::getSingleton(
     css::uno::Reference< css::uno::XComponentContext > const & context)
 {
     assert(context.is());
-    return theComponentsSingleton::get(context);
+    static Components singleton(context);
+    return singleton;
 }
 
 bool Components::allLocales(OUString const & locale) {
@@ -224,10 +216,9 @@ rtl::Reference< Node > Components::resolvePathRepresentation(
         pathRepresentation, canonicRepresentation, path, finalizedLayer);
 }
 
-rtl::Reference< Node > Components::getTemplate(
-    int layer, OUString const & fullName) const
+rtl::Reference< Node > Components::getTemplate(OUString const & fullName) const
 {
-    return data_.getTemplate(layer, fullName);
+    return data_.getTemplate(Data::NO_LAYER, fullName);
 }
 
 void Components::addRootAccess(rtl::Reference< RootAccess > const & access) {
@@ -243,19 +234,21 @@ void Components::initGlobalBroadcaster(
     rtl::Reference< RootAccess > const & exclude, Broadcaster * broadcaster)
 {
     //TODO: Iterate only over roots w/ listeners:
-    for (WeakRootSet::iterator i(roots_.begin()); i != roots_.end(); ++i) {
+    for (auto const& elemRoot : roots_)
+    {
         rtl::Reference< RootAccess > root;
-        if ((*i)->acquireCounting() > 1) {
-            root.set(*i); // must not throw
+        if (elemRoot->acquireCounting() > 1) {
+            root.set(elemRoot); // must not throw
         }
-        (*i)->releaseNondeleting();
+        elemRoot->releaseNondeleting();
         if (root.is()) {
             if (root != exclude) {
                 std::vector<OUString> path(root->getAbsolutePath());
                 Modifications::Node const * mods = &modifications.getRoot();
-                for (auto j(path.begin()); j != path.end(); ++j) {
+                for (auto const& pathElem : path)
+                {
                     Modifications::Node::Children::const_iterator k(
-                        mods->children.find(*j));
+                        mods->children.find(pathElem));
                     if (k == mods->children.end()) {
                         mods = nullptr;
                         break;
@@ -364,9 +357,10 @@ void Components::removeExtensionXcuFile(
             rtl::Reference< Node > parent;
             NodeMap const * map = &data_.getComponents();
             rtl::Reference< Node > node;
-            for (auto j(i->begin()); j != i->end(); ++j) {
+            for (auto const& j : *i)
+            {
                 parent = node;
-                node = map->findNode(Data::NO_LAYER, *j);
+                node = map->findNode(Data::NO_LAYER, j);
                 if (!node.is()) {
                     break;
                 }
@@ -404,8 +398,7 @@ void Components::insertModificationXcuFile(
     } catch (css::container::NoSuchElementException & e) {
         SAL_WARN(
             "configmgr",
-            "error inserting non-existing \"" << fileUri << "\": \""
-                << e.Message << '"');
+            "error inserting non-existing \"" << fileUri << "\": " << e);
     }
 }
 
@@ -433,15 +426,13 @@ css::beans::Optional< css::uno::Any > Components::getExternalValue(
             // installed:
             SAL_WARN(
                 "configmgr",
-                "createInstance(" << name << ") failed with \"" << e.Message
-                    << '"');
+                "createInstance(" << name << ") failed with " << e);
         }
         css::uno::Reference< css::beans::XPropertySet > propset;
         if (service.is()) {
             propset.set( service, css::uno::UNO_QUERY_THROW);
         }
-        j = externalServices_.insert(
-            ExternalServices::value_type(name, propset)).first;
+        j = externalServices_.emplace(name, propset).first;
     }
     css::beans::Optional< css::uno::Any > value;
     if (j->second.is()) {
@@ -456,8 +447,10 @@ css::beans::Optional< css::uno::Any > Components::getExternalValue(
             throw css::uno::RuntimeException(
                 "unknown external value descriptor ID: " + e.Message);
         } catch (css::lang::WrappedTargetException & e) {
-            throw css::uno::RuntimeException(
-                "cannot obtain external value: " + e.Message);
+            css::uno::Any anyEx = cppu::getCaughtException();
+            throw css::lang::WrappedTargetRuntimeException(
+                "cannot obtain external value: " + e.Message,
+                nullptr, anyEx );
         }
     }
     return value;
@@ -488,7 +481,7 @@ Components::Components(
         for (;; ++c) {
             if (c == conf.getLength() || conf[c] == ' ') {
                 throw css::uno::RuntimeException(
-                    "CONFIGURATION_LAYERS: missing \":\"");
+                    "CONFIGURATION_LAYERS: missing ':' in \"" + conf + "\"");
             }
             if (conf[c] == ':') {
                 break;
@@ -560,8 +553,9 @@ Components::Components(
             }
             OUString aTempFileURL;
             if (dumpWindowsRegistry(&aTempFileURL, eType)) {
-                parseFileLeniently(&parseXcuFile, aTempFileURL, layer, 0, 0, 0);
-                osl::File::remove(aTempFileURL);
+                parseFileLeniently(&parseXcuFile, aTempFileURL, layer, nullptr, nullptr, nullptr);
+                if (!getenv("SAL_CONFIG_WINREG_RETAIN_TMP"))
+                    osl::File::remove(aTempFileURL);
             }
             ++layer; //TODO: overflow
 #endif
@@ -613,9 +607,38 @@ Components::Components(
 
 Components::~Components()
 {
-    flushModifications();
-    for (WeakRootSet::iterator i(roots_.begin()); i != roots_.end(); ++i) {
-        (*i)->setAlive(false);
+    // get flag if _exit was already called which is a sign to not secure user config.
+    // this is used for win only currently where calling _exit() unfortunately still
+    // calls destructors (what is not wanted). May be needed for other systems, too
+    // (unknown yet) but can do no harm
+    const bool bExitWasCalled(comphelper::BackupFileHelper::getExitWasCalled());
+
+#ifndef WNT
+    // we can add a SAL_WARN here for other systems where the destructor gets called after
+    // an _exit() call. Still safe - the getExitWasCalled() is used, but a hint that _exit
+    // behaves different on a system
+    SAL_WARN_IF(bExitWasCalled, "configmgr", "Components::~Components() called after _exit() call");
+#endif
+
+    if (bExitWasCalled)
+    {
+        // do not write, re-join threads
+        osl::MutexGuard g(*lock_);
+
+        if (writeThread_.is())
+        {
+            writeThread_->join();
+        }
+    }
+    else
+    {
+        // write changes
+        flushModifications();
+    }
+
+    for (auto const& rootElem : roots_)
+    {
+        rootElem->setAlive(false);
     }
 }
 
@@ -634,7 +657,7 @@ void Components::parseFileLeniently(
         // starting:
         SAL_WARN(
             "configmgr",
-            "error reading \"" << url << "\": \"" << e.Message << '"');
+            "error reading \"" << url << "\": " << e);
     }
 }
 
@@ -650,7 +673,7 @@ void Components::parseFiles(
         if (!recursive) {
             return;
         }
-        // fall through
+        [[fallthrough]];
     default:
         throw css::uno::RuntimeException(
             "cannot open directory " + url);
@@ -681,6 +704,10 @@ void Components::parseFiles(
                     parseFileLeniently(
                         parseFile, stat.getFileURL(), layer, nullptr, nullptr, nullptr);
                 } catch (css::container::NoSuchElementException & e) {
+                    if (stat.getFileType() == osl::FileStatus::Link) {
+                        SAL_WARN("configmgr", "dangling link <" << stat.getFileURL() << ">");
+                        continue;
+                    }
                     throw css::uno::RuntimeException(
                         "stat'ed file does not exist: " + e.Message);
                 }
@@ -703,8 +730,7 @@ void Components::parseFileList(
             try {
                 parseFileLeniently(parseFile, url, layer, nullptr, nullptr, adds);
             } catch (css::container::NoSuchElementException & e) {
-                SAL_WARN(
-                    "configmgr", "file does not exist: \"" << e.Message << '"');
+                SAL_WARN("configmgr", "file does not exist: " << e);
                 if (adds != nullptr) {
                     data_.removeExtensionXcuAdditions(url);
                 }
@@ -727,7 +753,7 @@ void Components::parseXcdFiles(int layer, OUString const & url) {
         throw css::uno::RuntimeException(
             "cannot open directory " + url);
     }
-    UnresolvedList unres;
+    UnresolvedVector unres;
     std::set< OUString > existingDeps;
     std::set< OUString > processedDeps;
     for (;;) {
@@ -758,20 +784,24 @@ void Components::parseXcdFiles(int layer, OUString const & url) {
                         stat.getFileURL(),
                         new XcdParser(layer, processedDeps, data_));
                 } catch (css::container::NoSuchElementException & e) {
+                    if (stat.getFileType() == osl::FileStatus::Link) {
+                        SAL_WARN("configmgr", "dangling link <" << stat.getFileURL() << ">");
+                        continue;
+                    }
                     throw css::uno::RuntimeException(
                         "stat'ed file does not exist: " + e.Message);
                 }
                 if (manager->parse(nullptr)) {
                     processedDeps.insert(name);
                 } else {
-                    unres.push_back(UnresolvedListItem(name, manager));
+                    unres.emplace_back(name, manager);
                 }
             }
         }
     }
     while (!unres.empty()) {
         bool isResolved = false;
-        for (UnresolvedList::iterator i(unres.begin()); i != unres.end();) {
+        for (UnresolvedVector::iterator i(unres.begin()); i != unres.end();) {
             if (i->manager->parse(&existingDeps)) {
                 processedDeps.insert(i->name);
                 i = unres.erase(i);
@@ -808,7 +838,7 @@ void Components::parseXcsXcuIniLayer(
             case ':':
             case '\\':
                 prefix.append('\\');
-                // fall through
+                [[fallthrough]];
             default:
                 prefix.append(c);
             }
@@ -855,7 +885,7 @@ void Components::parseModificationLayer(int layer, OUString const & url) {
     }
 }
 
-int Components::getExtensionLayer(bool shared) {
+int Components::getExtensionLayer(bool shared) const {
     int layer = shared ? sharedExtensionLayer_ : userExtensionLayer_;
     if (layer == -1) {
         throw css::uno::RuntimeException(
